requirements.txt
# disco-py==0.0.12
# holster==latest
click==7.0
-e git+https://github.com/b1naryth1ef/holster.git#egg=holster
-e git+https://github.com/b1naryth1ef/disco.git@staging/v1.0.0#egg=disco-py
Flask==1.0.3
-e git+https://github.com/elderlabs/peewee.git@master#egg=peewee
telecom-py==0.0.4
requests-oauthlib==1.2.0
redis==3.2.1
regex==2019.2.3
PyYAML==5.1.1
ujson==1.35
oauth2client==4.1.3
oauthlib==3.0.1
psycopg2==2.8.3
psutil==5.6.3
Werkzeug==0.14.1
gevent-eventemitter==2.1
youtube_dl==2019.7.30